requirements.txt
pandas
numpy
scikit-learn
shap
fastapi
uvicorn[standard]
boto3
pydantic
joblib
pyyaml
python-multipart
ipykernel
python-box
ensure
openpyxl